added food models and migrations

+++ JWTAuthentication.WebApi/Models/FoodModels/Day.cs
@@ -4,11 +4,11 @@ namespace JWTAuthentication.WebApi.Models.FoodModels
 {
     public  class Day
     {
-        public byte Id { get; set; }
+        public int Id { get; set; }
         public string WeekDay { get; set; }
         public TimeSpan StartingHour { get; set; }
         public TimeSpan ClosingHour { get; set; }
-        public byte OperatingTimeId { get; set; }
+        public int OperatingTimeId { get; set; }
 
         public  OperatingTime OperatingTime { get; set; }
     }
